Clear tools config and options, add docs #4, add GUI #140
