+++ howard/main.py
@@ -28,7 +28,10 @@ from howard.tools.tools import *
 
 # Usage
 # python -m pip install -e .
+# howard --help
+# howard query --help
 # howard analysis --input=my.vcf.gz --output=my.output.vcf --annotations=my.annotations.vcf.gz --stats --overview_footer
+# howard gui
 # python -m howard.main --input=my.vcf.gz --output=my.output.vcf --annotations=my.annotations.vcf.gz --stats --overview_footer
 
 
@@ -39,123 +42,18 @@ def main() -> None:
     """
     It loads a VCF file in multiple format (VCF, parquet, DB), and process, query, export data
     """
+
+    # Config infos
+    setup_cfg = f'{main_folder}/../setup.cfg'
+    arguments_dict = {
+        "arguments": arguments,
+        "commands_arguments": commands_arguments,
+        "shared_arguments": shared_arguments 
+    }
     
-    # Config Parser
-    cf = ConfigParser()
-    cf.read(f'{main_folder}/../setup.cfg')
-    prog_name = cf['metadata']['name']
-    prog_version = cf['metadata']['version']
-    prog_description = cf['metadata']['description']
-    prog_long_description_content_type = cf['metadata']['long_description_content_type']
-    # prog_author = cf['metadata']['author']
-    # prog_author_email = cf['metadata']['author_email']
-
-    # Main parser
-    parser = argparse.ArgumentParser(
-        prog="howard",
-        description=f"""{prog_name.upper()}:{prog_version}\n"""
-                    f"""{prog_description}\n"""
-                    f"""{prog_long_description_content_type}""",
-        epilog="Usage examples:\n"
-            """   howard process --input=tests/data/example.vcf.gz --output=/tmp/example.annotated.vcf.gz --param=config/param.json \n"""
-            """   howard annotation --input=tests/data/example.vcf.gz --output=/tmp/example.howard.vcf.gz --annotations='tests/databases/annotations/hg19/dbnsfp42a.parquet,tests/databases/annotations/hg19/gnomad211_genome.parquet' \n"""
-            """   howard calculation --input=tests/data/example.full.vcf --output=/tmp/example.calculation.tsv --calculations='vartype' \n"""
-            """   howard prioritization --input=tests/data/example.vcf.gz --output=/tmp/example.prioritized.vcf.gz --prioritizations=config/prioritization_profiles.json --profiles='default,GERMLINE' \n"""
-            """   howard query --input=tests/data/example.vcf.gz --explode_infos --query='SELECT "#CHROM", POS, REF, ALT, "DP", "CLNSIG", sample2, sample3 FROM variants WHERE "DP" >= 50 OR "CLNSIG" NOT NULL ORDER BY "CLNSIG" DESC, "DP" DESC' \n"""
-            """   howard stats --input=tests/data/example.vcf.gz \n"""
-            """   howard convert --input=tests/data/example.vcf.gz --output=/tmp/example.tsv --explode_infos && cat /tmp/example.tsv \n"""
-            ,
-        formatter_class=argparse.RawTextHelpFormatter
-        )
-    
-    parser._optionals.title = "Shared arguments"
-
-    subparsers = parser.add_subparsers(title="Tools", dest='command')
-
-    options_md = ""
-    options_html = ""
-    
-    options_md += f"# HOWARD Help"
-    options_md += "\n\n"
-    options_md += parser.description.replace("\n","\n\n")
-    options_md += re.sub(r'> $',"",parser.epilog.replace("\n","\n\n").replace("   ","> "))
-    options_md += "\n\n"
-
-    options_html += f"<H1>HOWARD Help</h1>\n"
-    options_html += "<p>" + parser.description.replace("\n","<br>") + "</p>"
-    options_html += parser.epilog.replace("\n","<br>").replace("   ","&nbsp;&nbsp;&nbsp;")
-        
-
-    # Create commands arguments
-    for command in commands_arguments:
-        
-        command_description = commands_arguments[command].get("description","")
-        command_help = commands_arguments[command].get("help","")
-        command_epilog = commands_arguments[command].get("epilog","")
-        
-        command_parser = subparsers.add_parser(
-            command,
-            description = command_description,
-            help = command_help,
-            epilog = command_epilog,
-            formatter_class=argparse.RawTextHelpFormatter
-        )
-
-        # Markdown
-        options_md += f"## {command.upper()} tool\n"
-        options_md += command_description.replace("\n","\n\n")
-        options_md += "\n\n"
-        options_md += re.sub(r'> $',"",command_epilog.replace("\n","\n\n").replace("   ","> "))
-        options_md += "\n\n"
-
-        # HTML
-        options_html += f"<H2>{command.upper()}</H2>\n"
-        options_html += "<p>" + command_description.replace("\n","<br>") + "</p>"
-        options_html += command_epilog.replace("\n","<br>").replace("   ","&nbsp;&nbsp;&nbsp;")
-        
-        
-        # Main args
-        command_parser._optionals.title = "Options"
-        if "main" in commands_arguments[command]["groups"]:
-            group = "main"
-            options_md += f"### Main options\n"
-            options_html += f"<H3>Main options</H3>\n"
-            for arg in commands_arguments[command]["groups"][group]:
-                required = commands_arguments[command]["groups"][group][arg]
-                argument = get_argument(arguments=arguments.copy(), arg=arg, required=required)
-                command_parser.add_argument(f"--{arg}", **argument)
-                options_md += get_argument_to_mk(arg, argument)
-                options_html += get_argument_to_mk(arg, argument, mode="html")
-
-        for group in commands_arguments[command]["groups"]:
-            if group != "main":
-                options_md += f"### {group}\n"
-                options_html += f"<H3>{group}</H3>\n"
-                command_group = command_parser.add_argument_group(f"{group} options")
-                for arg in commands_arguments[command]["groups"][group]:
-                    required = commands_arguments[command]["groups"][group][arg]
-                    argument = get_argument(arguments=arguments.copy(), arg=arg, required=required)
-                    command_group.add_argument(f"--{arg}", **argument)
-                    options_md += get_argument_to_mk(arg, argument)
-                    options_html += get_argument_to_mk(arg, argument, mode="html")
-
-        # Shared arguments
-        shared_group = command_parser.add_argument_group('Shared options')
-        for arg in shared_arguments:
-            shared_group.add_argument(f"--{arg}", **get_argument(arguments=arguments, arg=arg, required=False))
-
-        #options_mk += "> " + re.sub(r'> $',"",command_epilog.replace("\n","\n\n> "))
-        options_md += "\n\n"
-
-
-    options_md += f"## Shared arguments\n"
-    options_html += f"<H2>Shared arguments</H2>\n".upper()
-    for arg in shared_arguments:
-        required = False
-        argument = get_argument(arguments=arguments.copy(), arg=arg, required=required)
-        options_md += get_argument_to_mk(arg, argument)
-        options_html += get_argument_to_mk(arg, argument, mode="html")
- 
+    # Generate parser
+    parser = argparse.ArgumentParser()
+    parser = help_generation(arguments_dict=arguments_dict, parser=parser, setup=setup_cfg, output_type="parser")
 
     # Parse args
     args, remaining = parser.parse_known_args()
@@ -175,6 +73,10 @@ def main() -> None:
     # log
     if "log" not in args:
         args.log = None
+
+    # Config infos
+    args.arguments_dict = arguments_dict
+    args.setup_cfg = setup_cfg
 
     # Logging
     set_log_level(args.verbosity, args.log)
@@ -246,26 +148,17 @@ def main() -> None:
     # Command eval
     if not args.command:
         parser.print_help()
-        return
-    elif args.command == "help":
-        if "help_md" in args and args.help_md is not None:
-            help_file = args.help_md.name
-            f = open(help_file, "w")
-            f.write(options_md)
-            f.close()
-        if "help_html" in args and args.help_html is not None:
-            help_file = args.help_html.name
-            f = open(help_file, "w")
-            f.write(options_html)
-            f.close()
-        if "help_md" not in args and "help_html" not in args:
-            parser.print_help()
-        return
     else:
+        if args.command == "gui" and not tool_gui_enable:
+            #from gooey import Gooey, GooeyParser
+            log.error("""HOWARD GUI disabled""")
+            log.error("""ModuleNotFoundError: No module named 'gooey'""")
+            log.error("""Install module 'gooey': pip install gooey""")
+            log.error("""Or install requirements: pip install -r requirements-gui.txt""")
+            raise ValueError("""HOWARD GUI disabled""")
         command_function = commands_arguments[args.command]["function"]
         log.debug(f"Command/Tool: {command_function}")
         eval(f"{command_function}(args)")
-
 
 if __name__ == '__main__':
     main()
