+++ requirements-gui.txt
@@ -20,11 +20,12 @@ mgzip==0.2.1
 pysam==0.21.0
 jproperties==2.1.1
 psutil==5.9.6
-wxpython>=4.1.0
+markdown==3.5.2
+tabulate==0.9.0
 Pillow>=4.3.0
-psutil>=5.4.2
 colored>=1.3.93
 pygtrie>=2.3.3
 re-wx>=0.0.2
 typing-extensions==3.10.0.2
 mypy-extensions==0.4.3
+wxpython>=4.1.0
