+++ requirements.txt
@@ -20,3 +20,5 @@ mgzip==0.2.1
 pysam==0.21.0
 jproperties==2.1.1
 psutil==5.9.6
+markdown==3.5.2
+tabulate==0.9.0
