--- setup-gui.cfg
@@ -1,46 +1,0 @@
-[metadata]
-name = howard-gui
-version = 1.0.0
-description = A Python project
-author = Antony Le Bechec
-author_email = antony.lebechec@chru-strasbourg.fr
-
-[options]
-python_requires = >=3.10
-packages = find:
-install_requires =
-    pandas==2.1.0
-    duckdb==0.9.1
-    pyarrow==13.0.0
-    bio==1.5.9
-    pyvcf3==1.0.3
-    dask==2023.9.0
-    pytest==7.4.1
-    coverage==7.3.1
-    fastparquet==2023.8.0
-    polars==0.19.2
-    genomepy==0.16.1
-    flake8==6.1.0
-    lazy==1.5
-    pynose==1.4.8
-    pyfaidx==0.7.2.2
-    multiprocesspandas==1.1.5
-    bgzip==0.4.0
-    pgzip==0.3.5
-    mgzip==0.2.1
-    pysam==0.21.0
-    jproperties==2.1.1
-    psutil==5.9.6
-    wxpython>=4.1.0
-    Pillow>=4.3.0
-    psutil>=5.4.2
-    colored>=1.3.93
-    pygtrie>=2.3.3
-    re-wx>=0.0.2
-    typing-extensions==3.10.0.2
-    mypy-extensions==0.4.3
-    
-
-[options.entry_points]
-console_scripts =
-    howard = howard.gui:main
