+++ setup.cfg
@@ -2,7 +2,7 @@
 name = howard
 version = 1.0.0
 description = Highly Open and Valuable tool for Variant Annotation & Ranking for Discovery
-long_description_content_type = HOWARD annotates and prioritizes genetic variations, calculates and normalizes annotations, translates vcf format and generates variants statistics
+long_description_content_type = HOWARD annotates and prioritizes genetic variations, calculates and normalizes annotations, convert on multiple formats, query variations and generates statistics
 keywords = VCF variant annotation ranking
 author = Antony Le Bechec
 author_email = bioinfo@chru-strasbourg.fr
@@ -34,6 +34,7 @@ install_requires =
     jproperties==2.1.1
     psutil==5.9.6
     markdown==3.5.2
+    tabulate==0.9.0
     
 
 [options.entry_points]
